+++ tests/utils/rule.py
@@ -1,6 +1,6 @@
 from copy import deepcopy
 from random import choice, randint
-from starkinfra import IssuingRule
+from starkinfra import IssuingRule, MerchantCountry, CardMethod, MerchantCategory
 
 
 example_rule = IssuingRule(
@@ -17,5 +17,11 @@ def generateExampleRuleJson(n=1):
         example_rule.interval = choice(["day", "week", "month", "instant"])
         example_rule.amount = randint(1000, 100000)
         example_rule.currency_code = choice(["BRL", "USD"])
+        example_rule.countries = [MerchantCountry(code=choice(["BRA", "USA"]))]
+        example_rule.methods = [CardMethod(code=choice(["manual", "server", "token"]))]
+        example_rule.categories = [choice([
+            MerchantCategory(code=choice(["fastFoodRestaurants", "veterinaryServices"])),
+            MerchantCategory(type=choice(["pets", "food"]))
+        ])]
         rules.append(deepcopy(example_rule))
     return rules